requirements.txt
numpy
opencv-python
Pillow
matplotlib
scipy
pyyaml
ultralytics
tqdmZ